refactor: move product list page to pages directory
Moves ProductList page component to pages folder as it is a page
